requirements.txt
matplotlib
torch
transformers
numpy
pillow
pandas
beautifulsoup4
requestsls
uvicorn
streamlit